I restructured your code so that the lines are not longer than 80 characters (generally, lines shouldn't be longer than that).
I added an if __name__ == "__main__": statement, because if someone where to import your program, they would be forced to have the code run once, even if they only implement the function.

I also refractored someone your variables to be more pythonic, i.e. using underscores instead of camel case.

Lastly, I used the getpass method so that the second player cannot see what you are typing on the screen. You could use something like screen flush to achieve the same effect, but it would be a bit more code and more time-intensive. So, I opted for the easiser function